Add files via upload
Baccarat is a card game played at casinos. It is a comparing card game played between two hands, the "player" and the "banker". Each baccarat coup (round of play) has three possible outcomes: "player" (player has the higher score), "banker", and "tie".

Tableau of drawing rules
If neither the player nor the banker is dealt a total of 8 or 9 in the first two cards (known as a "natural"), the tableau is consulted, first for the player's rules, then the banker's.

Player's rule
If the player has an initial total of 5 or less, they draw a third card. If the player has an initial total of 6 or 7, they stand.
Banker's rule
If the player stood pat (i.e. has only two cards), the banker regards only their own hand and acts according to the same rule as the player, i.e. the banker draws a third card with hands 5 or less and stands with 6 or 7.
If the player drew a third card, the banker acts according to the following more complex rules:
If the banker total is 2 or less, they draw a third card, regardless of what the player's third card is.
If the banker total is 3, they draw a third card unless the player's third card is an 8.
If the banker total is 4, they draw a third card if the player's third card is 2, 3, 4, 5, 6, 7.
If the banker total is 5, they draw a third card if the player's third card is 4, 5, 6, or 7.
If the banker total is 6, they draw a third card if the player's third card is a 6 or 7.
If the banker total is 7, they stand.


banker
total                       player's third card

0	1	2	3	4	5	6	7	8	9
<3	H	H	H	H	H	H	H	H	H	H
3	H	H	H	H	H	H	H	H	S	H
4	S	S	H	H	H	H	H	H	S	S
5	S	S	S	S	H	H	H	H	S	S
6	S	S	S	S	S	S	H	H	S	S
>6	S	S	S	S	S	S	S	S	S	S

The croupier will deal the cards according to the tableau and the croupier will announce the winning hand, either the player or the banker. Losing bets will be collected and the winning bets will be paid according to the rules of the house. Usually, 1-to-1 even money will be paid on player bets and 19-to-20 on banker bets (even money with "5% commission to the house on the win").

Should both the player and banker have the same value at the end of the deal the croupier shall announce "tie bets win." All tie bets will be paid at 8-to-1 odds and all bets on player or banker remain in place and active for the next game (the customer may or may not be able to retract these bets depending on casino rules).